+++ raster/src/test/scala/geotrellis/raster/io/geotiff/GeoTiffMultibandTileSpec.scala
@@ -720,6 +720,57 @@ class GeoTiffMultibandTileSpec extends FunSpec
     }
   }
 
+  describe("GeoTiffMultibandTile crop") {
+    val pixelStripedGeoTiff = GeoTiffReader.readMultiband(geoTiffPath("3bands/int32/3bands-striped-pixel.tif"))
+    val pixelTiledGeoTiff = GeoTiffReader.readMultiband(geoTiffPath("3bands/int32/3bands-tiled-pixel.tif"))
+
+    val bandStripedGeoTiff = GeoTiffReader.readMultiband(geoTiffPath("3bands/int32/3bands-striped-band.tif"))
+    val bandTiledGeoTiff = GeoTiffReader.readMultiband(geoTiffPath("3bands/int32/3bands-tiled-band.tif"))
+
+    val pixelStripedRasterExtent = RasterExtent(pixelStripedGeoTiff.extent, pixelStripedGeoTiff.cols, pixelStripedGeoTiff.rows)
+    val bandStripedRasterExtent = RasterExtent(bandStripedGeoTiff.extent, bandStripedGeoTiff.cols, bandStripedGeoTiff.rows)
+
+    it("should have the correct number of subset bands - pixel") {
+      val cropped = pixelStripedGeoTiff.tile.cropBands(pixelStripedRasterExtent.gridBounds, Array(1, 0))
+
+      cropped.bands.size should be (2)
+    }
+
+    it("should have the correct number of subset bands - band") {
+      val cropped = bandStripedGeoTiff.tile.cropBands(bandStripedRasterExtent.gridBounds, Array(1))
+
+      cropped.bands.size should be (1)
+    }
+
+    it("should have the crop the correct area - pixel striped") {
+      val actual = pixelStripedGeoTiff.tile.cropBands(pixelStripedRasterExtent.gridBounds, Array(1, 0, 2))
+      val expected = pixelStripedGeoTiff.crop(pixelStripedRasterExtent.gridBounds).tile.subsetBands(1, 0, 2)
+
+      actual should be (expected)
+    }
+
+    it("should have the crop the correct area - pixel tiled") {
+      val actual = pixelTiledGeoTiff.tile.cropBands(pixelStripedRasterExtent.gridBounds, Array(1, 0, 2))
+      val expected = pixelTiledGeoTiff.crop(pixelStripedRasterExtent.gridBounds).tile.subsetBands(1, 0, 2)
+
+      actual should be (expected)
+    }
+
+    it("should have the crop the correct area - band striped") {
+      val actual = bandStripedGeoTiff.tile.cropBands(bandStripedRasterExtent.gridBounds, Array(1, 2, 0))
+      val expected = bandStripedGeoTiff.crop(bandStripedRasterExtent.gridBounds).tile.subsetBands(1, 2, 0)
+
+      actual should be (expected)
+    }
+
+    it("should have the crop the correct area - band tiled") {
+      val actual = bandTiledGeoTiff.tile.cropBands(bandStripedRasterExtent.gridBounds, Array(1, 2, 0))
+      val expected = bandTiledGeoTiff.crop(bandStripedRasterExtent.gridBounds).tile.subsetBands(1, 2, 0)
+
+      actual should be (expected)
+    }
+  }
+
   describe("GeoTiffMultibandTile streaming read") {
     it("reads over-buffered windows"){
       val path = geoTiffPath("3bands/int32/3bands-striped-pixel.tif")